api_yamdb/requirements.txt
 asgiref==3.2.10
 Django==2.2.16
 pytz==2021.3
 sqlparse==0.4.2
 pytest==6.2.4
 pytest-pythonpath==0.7.3
 pytest-django==4.4.0
 django-extra-fields==3.0.2
 django-filter==2.4.0
 djangorestframework==3.12.4
 djangorestframework-simplejwt==4.7.2
 djoser==2.1.0
 drf-spectacular==0.18.0
 flake8==3.9.2
 flake8-broken-line==0.3.0
 flake8-isort==4.0.0
 flake8-plugin-utils==1.3.2
 flake8-polyfill==1.0.2
 flake8-return==1.1.3
 gunicorn==20.1.0
 isort==5.9.3
 itypes==1.2.0
 Pillow==8.3.1
 psycopg2-binary==2.8.6
 pycodestyle==2.7.0
 pycparser==2.20
 pyflakes==2.3.1
 PyJWT==2.1.0
 python-dotenv==0.19.2
 requests==2.26.0
 testfixtures==6.18.1